requirements.txt
alembic==1.7.1








elasticsearch==7.13.3

Flask==2.0.1


Flask-HTTPAuth==4.4.0


Flask-Migrate==3.0.1
Flask-Moment==1.0.1



httpie==2.4.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
langdetect==1.0.9
Mako==1.1.4
MarkupSafe==2.0.1

PyJWT==2.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0

pytz==2021.1

requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.20





# requirements for Heroku
#psycopg2==2.9.1
#gunicorn==20.1.0


ariadne==0.13.0
arrow==1.1.1
astroid==2.6.2
async-generator==1.10

attrs==21.2.0
autopep8==1.5.7

Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
bidict==0.21.2
billiard==3.6.4.0
binaryornot==0.4.4
bitarray==2.3.2
bleach==3.3.0
blinker==1.4
cachelib==0.3.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==1.6.0
clr==1.0.3
colorama==0.4.4
ConfigArgParse==1.5.2
configparser==5.0.2
cookiecutter==1.7.3
cryptography==3.4.7
Cython==0.29.24
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
direct-redis==0.3.1
distlib==0.3.2
dnspython==2.1.0
docutils==0.17.1
dominate==2.6.0
email-validator==1.1.3
entrypoints==0.3
eventlet==0.32.0
filelock==3.0.12
flake8==2.3.0
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
flask_moment
Flask-WTF==0.15.1
flat-table==1.1.1
gevent==21.8.0
greenlet==1.1.0
gssapi==1.6.14
gunicorn==20.1.0
jwt==1.2.0

numpy==1.21.2
numpydoc==1.1.0

pandas==1.3.2

parso==0.5.2
pep8==1.7.1

pipenv==2021.5.29

pyflakes==2.1.1
Pygments==2.9.0
pylint==2.9.3



python-decouple==3.4
python-editor==1.0.4
python-engineio==3.13.2
python-jsonrpc-server==0.4.0
python-language-server==0.31.10
python-slugify==5.0.2
python-socketio==4.6.0
redis==3.5.3
redistimeseries==1.4.3
rq==1.10.0
sse==1.2
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
visitor==0.1.3
walrus==0.8.2
watchdog==2.1.3
webencodings==0.5.1
Werkzeug==2.0.1
WTForms==2.3.3
